split out event properties, trace context, and event names into separate files consistently
