+++ src/nft/properties.ts
@@ -1,0 +1,9 @@
+/**
+ * Known Filter Types for NFTs
+ */
+// eslint-disable-next-line @typescript-eslint/naming-convention
+export enum NFTFilterTypes {
+  MARKETPLACE = 'Marketplace',
+  PRICE_RANGE = 'Price Range',
+  TRAIT = 'Trait',
+}
