+++ src/swap/properties.ts
@@ -1,0 +1,4 @@
+export enum SwapPriceUpdateUserResponse {
+  ACCEPTED = 'Accepted',
+  REJECTED = 'Rejected',
+}
